:recycle: refactor(cleanup): 유틸 함수를 한곳에 모아놓는다
양이 얼마 없기 때문에 한곳에 모아놔도 괜찮다고 생각했다
