--- src/js/utils/dom.js
@@ -1,7 +1,0 @@
-export const $ = (selector, parentNode = document) => parentNode.querySelector(selector);
-
-export const removeChildren = parentNode => {
-  while (parentNode.firstChild) {
-    parentNode.removeChild(parentNode.firstChild);
-  }
-};
